Fix/add statistics of all the heaps
The statistics in the binary heap was not initialized correctly, nor was it being updated; this has now been fixed. Moreover, statistics have been added to the binary heap with chaining class (monitoring the largest size the heap reaches during the process), as well as to the funnel heap (with and without merging), where the number of sweeps to each link is monitored.