Fix bugs affecting both Funnel Heap versions that resulted in incorrect output or segmentation fault
The incorrect output of the Funnel Heap with Merging is mostly a result of not knowing when to stop refilling a buffer (in the refillBuffer method). In the current version, the refillBuffer method stops when either the buffer has been refilled with the amount of elements specified in a parameter, or if adding the next element would break the invariant between the parent and its children. The latter case was causing the problem; assume we are refilling the buffer A_1 with an element of degree 10, we were checking if the buffers B_1 and A_2 had any elements larger than 10. However, if for instance B_1 was empty, we were NOT checking if the children of B_1 had any elements larger than 10 (which would also break the invariant). This bug was fixed, and till now no other (unresolved) bugs have been found.